+++ amd/src/editor.js
@@ -15,7 +15,7 @@
 
 /**
  * @module mod_onlyofficeeditor/editor
- * @copyright  2024 Ascensio System SIA <integration@onlyoffice.com>
+ * @copyright  2025 Ascensio System SIA <integration@onlyoffice.com>
  * @copyright  based on work by 2018 Olumuyiwa Taiwo <muyi.taiwo@logicexpertise.com>
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  **/
@@ -154,12 +154,8 @@ define(['jquery'], function($) {
                 var docEditor = null;
                 var config = data.config;
 
-                var favicon = config.documentType;
-                if (config.fileType === 'docxf' || config.fileType === 'oform') {
-                    favicon = config.fileType;
-                }
                 document.head.innerHTML += '<link type="image/x-icon" rel="icon" href="/mod/onlyofficeeditor/pix/'
-                    + favicon + '.ico" />';
+                    + config.documentType + '.ico" />';
 
                 var canAddInstance = data.addinstance;
 
@@ -240,9 +236,9 @@ define(['jquery'], function($) {
                     config.events.onRequestSendNotify = onRequestSendNotify;
                 }
 
-                if ((config.document.fileType === "docxf" || config.document.fileType === "oform")
+                if ((config.document.fileType === "pdf")
                     // eslint-disable-next-line no-undef
-                    && DocsAPI.DocEditor.version().split(".")[0] < 7) {
+                    && DocsAPI.DocEditor.version().split(".")[0] < 8) {
                     displayNotification('oldversion', 'error');
                 } else {
                     // eslint-disable-next-line no-undef