+++ classes/document.php
@@ -19,7 +19,7 @@
  *
  * @package     mod_onlyofficeeditor
  * @subpackage
- * @copyright   2024 Ascensio System SIA <integration@onlyoffice.com>
+ * @copyright   2025 Ascensio System SIA <integration@onlyoffice.com>
  * @copyright   based on work by 2018 Olumuyiwa <muyi.taiwo@logicexpertise.com>
  * @license     http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
@@ -31,7 +31,7 @@ namespace mod_onlyofficeeditor;
  *
  * @package     mod_onlyofficeeditor
  * @subpackage
- * @copyright   2024 Ascensio System SIA <integration@onlyoffice.com>
+ * @copyright   2025 Ascensio System SIA <integration@onlyoffice.com>
  * @copyright   based on work by 2018 Olumuyiwa <muyi.taiwo@logicexpertise.com>
  * @license     http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
@@ -79,6 +79,7 @@ class document {
         $permissions['print'] = empty($permissions['print']) ? $canmanage : true;
         $permissions['download'] = empty($permissions['download']) ? $canmanage : true;
         $permissions['protect'] = $canmanage || empty($permissions['protect']);
+        $permissions['chat'] = get_config('onlyofficeeditor')->editor_view_chat == 1 && !is_guest($context);
 
         $permissions['edit'] = $canedit;
         $permissions['review'] = $canedit;