+++ classes/onlyoffice_file_utility.php
@@ -19,7 +19,7 @@
  *
  * @package     mod_onlyofficeeditor
  * @subpackage
- * @copyright   2024 Ascensio System SIA <integration@onlyoffice.com>
+ * @copyright   2025 Ascensio System SIA <integration@onlyoffice.com>
  * @license        http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
@@ -30,7 +30,7 @@ namespace mod_onlyofficeeditor;
  *
  * @package     mod_onlyofficeeditor
  * @subpackage
- * @copyright   2024 Ascensio System SIA <integration@onlyoffice.com>
+ * @copyright   2025 Ascensio System SIA <integration@onlyoffice.com>
  * @license        http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 class onlyoffice_file_utility {
@@ -40,9 +40,9 @@ class onlyoffice_file_utility {
      * @return string[] Accepted extensions of spreadsheet files
      */
     public static function get_accepted_spreadsheet_formats() {
-        return array('.xls', '.xlsx', '.xlsm',
+        return ['.xls', '.xlsx', '.xlsm',
             '.xlt', '.xltx', '.xltm',
-            '.ods', '.fods', '.ots', '.csv');
+            '.ods', '.fods', '.ots', '.csv'];
     }
 
     /**
@@ -50,11 +50,11 @@ class onlyoffice_file_utility {
      * @return string[] Accepted extensions of document files.
      */
     public static function get_accepted_document_formats() {
-        return array('.doc', '.docx', '.docm',
+        return ['.doc', '.docx', '.docm',
             '.dot', '.dotx', '.dotm',
             '.odt', '.fodt', '.ott', '.rtf', '.txt',
             '.html', '.htm', '.mht', '.xml',
-            '.pdf', '.djvu', '.fb2', '.epub', '.xps', '.oxps', '.oform', '.docxf');
+            '.fb2', '.epub'];
     }
 
     /**
@@ -62,10 +62,53 @@ class onlyoffice_file_utility {
      * @return string[] Accepted extensions of presentation files.
      */
     public static function get_accepted_presentation_formats() {
-        return array('.pps', '.ppsx', '.ppsm',
+        return ['.pps', '.ppsx', '.ppsm',
             '.ppt', '.pptx', '.pptm',
             '.pot', '.potx', '.potm',
-            '.odp', '.fodp', '.otp');
+            '.odp', '.fodp', '.otp'];
+    }
+
+    /**
+     * Get accepted presentation extensions.
+     * @return string[] Accepted extensions of presentation files.
+     */
+    public static function get_accepted_pdf_formats() {
+        return ['.djvu', '.pdf', '.oform', '.docxf', '.xps', '.oxps'];
+    }
+
+    /**
+     * Return all the accepted formats
+     *
+     * @return array formats
+     */
+    public static function get_accepted_formats() {
+        $formats = array_unique(array_merge(
+            self::get_accepted_spreadsheet_formats(),
+            self::get_accepted_presentation_formats(),
+            self::get_accepted_document_formats(),
+            self::get_accepted_pdf_formats(),
+        ));
+
+        return $formats;
+    }
+
+
+    /**
+     * Determine if the format is supported by onlyoffice editor
+     *
+     * @param string $extension
+     * @return bool
+     */
+    public static function is_format_supported(string $extension) {
+        $formats = static::get_accepted_formats();
+
+        foreach ($formats as $format) {
+            if ($format === ".$extension") {
+                return true;
+            }
+        }
+
+        return false;
     }
 
     /**
@@ -83,6 +126,9 @@ class onlyoffice_file_utility {
         if (in_array($ext, self::get_accepted_presentation_formats())) {
             return 'slide';
         }
+        if (in_array($ext, self::get_accepted_pdf_formats())) {
+            return 'pdf';
+        }
         return 'word';
     }
 
@@ -91,7 +137,7 @@ class onlyoffice_file_utility {
      * @return string[] Editable extensions.
      */
     public static function get_editable_extensions() {
-        return array('.docx', '.xlsx', '.pptx', '.docxf', '.oform', '.pdf');
+        return ['.docx', '.xlsx', '.pptx', '.docxf', '.oform', '.pdf'];
     }
 
 }