+++ db/events.php
@@ -19,15 +19,15 @@
  *
  * @package     mod_onlyofficeeditor
  * @subpackage
- * @copyright   2024 Ascensio System SIA <integration@onlyoffice.com>
+ * @copyright   2025 Ascensio System SIA <integration@onlyoffice.com>
  * @license     http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 defined('MOODLE_INTERNAL') || die();
 
-$observers = array(
-    array(
+$observers = [
+    [
         'eventname' => '\core\event\user_loggedin',
         'callback'  => 'onlyofficeeditor_login_handler',
-        'includefile' => '/mod/onlyofficeeditor/lib.php'
-    )
-);
+        'includefile' => '/mod/onlyofficeeditor/lib.php',
+    ],
+];