+++ lib.php
@@ -25,7 +25,7 @@
  * Moodle is performing actions across all modules.
  *
  * @package    mod_onlyofficeeditor
- * @copyright  2024 Ascensio System SIA <integration@onlyoffice.com>
+ * @copyright  2025 Ascensio System SIA <integration@onlyoffice.com>
  * @copyright  based on work by 2018 Olumuyiwa Taiwo <muyi.taiwo@logicexpertise.com>
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
@@ -63,10 +63,10 @@ function onlyofficeeditor_supports($feature) {
  * of the new instance.
  *
  * @param stdClass $data Submitted data from the form in mod_form.php
- * @param mod_onlyofficeeditor_mod_form $mform The form instance itself (if needed)
+ * @param mod_onlyofficeeditor_mod_form|null $mform The form instance itself (if needed)
  * @return int The id of the newly inserted ONLYOFFICE record
  */
-function onlyofficeeditor_add_instance(stdClass $data, mod_onlyofficeeditor_mod_form $mform = null) {
+function onlyofficeeditor_add_instance(stdClass $data, ?mod_onlyofficeeditor_mod_form $mform = null) {
     global $CFG, $DB;
 
     $cmid = $data->coursemodule;
@@ -81,7 +81,7 @@ function onlyofficeeditor_add_instance(stdClass $data, mod_onlyofficeeditor_mod_
         $records = $DB->get_records('files', [
             'itemid' => $data->file,
             'component' => 'mod_onlyofficeeditor',
-            'filearea' => 'content'
+            'filearea' => 'content',
         ]);
         foreach ($records as $record) {
             $record->contextid = context_module::instance($cmid)->id;
@@ -93,7 +93,7 @@ function onlyofficeeditor_add_instance(stdClass $data, mod_onlyofficeeditor_mod_
     $data->id = $DB->insert_record('onlyofficeeditor', $data);
 
     // We need to use context now, so we need to make sure all needed info is already in db.
-    $DB->set_field('course_modules', 'instance', $data->id, array('id' => $cmid));
+    $DB->set_field('course_modules', 'instance', $data->id, ['id' => $cmid]);
 
     $completiontimeexpected = !empty($data->completionexpected) ? $data->completionexpected : null;
     \core_completion\api::update_completion_date_event($cmid, 'onlyofficeeditor', $data->id, $completiontimeexpected);
@@ -109,10 +109,10 @@ function onlyofficeeditor_add_instance(stdClass $data, mod_onlyofficeeditor_mod_
  * will update an existing instance with new data.
  *
  * @param stdClass $data An object from the form in mod_form.php
- * @param mod_onlyofficeeditor_mod_form $mform The form instance itself (if needed)
+ * @param mod_onlyofficeeditor_mod_form|null $mform The form instance itself (if needed)
  * @return boolean Success/Fail
  */
-function onlyofficeeditor_update_instance(stdClass $data, mod_onlyofficeeditor_mod_form $mform = null) {
+function onlyofficeeditor_update_instance(stdClass $data, ?mod_onlyofficeeditor_mod_form $mform = null) {
     global $CFG, $DB;
 
     $data->timemodified = time();
@@ -125,6 +125,12 @@ function onlyofficeeditor_update_instance(stdClass $data, mod_onlyofficeeditor_m
     \core_completion\api::update_completion_date_event($data->coursemodule, 'onlyofficeeditor', $data->id, $completiontimeexpected);
 
     $result = $DB->update_record('onlyofficeeditor', $data);
+
+    if ($data->file) {
+        $modinfo = get_fast_modinfo($data->course);
+        $cm = $modinfo->get_cm($data->coursemodule);
+        \mod_onlyofficeeditor\document::set_key($cm);
+    }
 
     return $result;
 }
@@ -142,14 +148,14 @@ function onlyofficeeditor_update_instance(stdClass $data, mod_onlyofficeeditor_m
 function onlyofficeeditor_delete_instance($id) {
     global $DB;
 
-    if (!$onlyoffice = $DB->get_record('onlyofficeeditor', array('id' => $id))) {
+    if (!$onlyoffice = $DB->get_record('onlyofficeeditor', ['id' => $id])) {
         return false;
     }
 
     $cm = get_coursemodule_from_instance('onlyofficeeditor', $id);
     \core_completion\api::update_completion_date_event($cm->id, 'onlyofficeeditor', $id, null);
 
-    $DB->delete_records('onlyofficeeditor', array('id' => $onlyoffice->id));
+    $DB->delete_records('onlyofficeeditor', ['id' => $onlyoffice->id]);
 
     return true;
 }
@@ -171,7 +177,7 @@ function onlyofficeeditor_get_coursemodule_info($coursemodule) {
 
     $context = \context_module::instance($coursemodule->id);
 
-    if (!$onlyoffice = $DB->get_record('onlyofficeeditor', array('id' => $coursemodule->instance),
+    if (!$onlyoffice = $DB->get_record('onlyofficeeditor', ['id' => $coursemodule->instance],
         'id, name, display, displayoptions, intro, introformat')) {
         return null;
     }
@@ -197,13 +203,13 @@ function onlyofficeeditor_get_coursemodule_info($coursemodule) {
 /**
  * Called when viewing course page. Shows extra details after the link if
  * enabled.
- * @todo Custom module instance display, similar to https://api.onlyoffice.com/editors/alfresco
+ * @todo Custom module instance display
  * @param cm_info $cm Course module information
  */
 function onlyofficeeditor_cm_info_view(cm_info $cm) {
     global $OUTPUT;
-    $icon = $OUTPUT->pix_icon('icon', get_string('onlyofficeactivityicon', 'onlyofficeeditor'), 'onlyofficeeditor',
-        array('class' => 'onlyofficeactivityicon'));
+    $icon = $OUTPUT->pix_icon('monologo', get_string('onlyofficeactivityicon', 'onlyofficeeditor'), 'onlyofficeeditor',
+        ['class' => 'onlyofficeactivityicon']);
 }
 
 /**
@@ -211,7 +217,7 @@ function onlyofficeeditor_cm_info_view(cm_info $cm) {
  *
  * This function is called from cm_info when displaying the module.
  *
- * @todo Custom module instance display, similar to https://api.onlyoffice.com/editors/alfresco
+ * @todo Custom module instance display
  * @param cm_info $cm
  */
 function onlyofficeeditor_cm_info_dynamic(cm_info $cm) {
@@ -310,7 +316,7 @@ function onlyofficeeditor_print_recent_mod_activity($activity, $courseid, $detai
  * @return array
  */
 function onlyofficeeditor_get_extra_capabilities() {
-    return array();
+    return [];
 }
 
 /* File API */
@@ -327,7 +333,7 @@ function onlyofficeeditor_get_extra_capabilities() {
  * @return array of [(string)filearea] => (string)description
  */
 function onlyofficeeditor_get_file_areas($course, $cm, $context) {
-    return array();
+    return [];
 }
 
 /**
@@ -364,7 +370,7 @@ function onlyofficeeditor_get_file_info($browser, $areas, $course, $cm, $context
  * @param bool $forcedownload whether or not force download
  * @param array $options additional options affecting the file serving
  */
-function onlyofficeeditor_pluginfile($course, $cm, $context, $filearea, array $args, $forcedownload, array $options = array()) {
+function onlyofficeeditor_pluginfile($course, $cm, $context, $filearea, array $args, $forcedownload, array $options = []) {
 
     $doc = required_param('doc', PARAM_RAW);
 
@@ -377,7 +383,8 @@ function onlyofficeeditor_pluginfile($course, $cm, $context, $filearea, array $a
     $modconfig = get_config('onlyofficeeditor');
     if (!empty($modconfig->documentserversecret)) {
         $jwtheader = !empty($modconfig->jwtheader) ? $modconfig->jwtheader : 'Authorization';
-        $token = substr(getallheaders()[$jwtheader], strlen('Bearer '));
+        $headers = array_change_key_case(getallheaders(), CASE_LOWER);
+        $token = substr($headers[strtolower($jwtheader)], strlen('Bearer '));
         try {
             $decodedheader = \mod_onlyofficeeditor\jwt_wrapper::decode($token, $modconfig->documentserversecret);
         } catch (\UnexpectedValueException $e) {
@@ -391,8 +398,11 @@ function onlyofficeeditor_pluginfile($course, $cm, $context, $filearea, array $a
     if (count($files) >= 1) {
         $file = reset($files);
         if ($hash->contenthash == $file->get_contenthash() && (is_enrolled($context, $hash->userid, '', true)
-                || has_any_capability(['moodle/course:manageactivities', 'mod/onlyofficeeditor:editdocument'],
-                    $context, $hash->userid))) {
+                || has_any_capability([
+                    'moodle/course:manageactivities',
+                    'mod/onlyofficeeditor:editdocument',
+                    'mod/onlyofficeeditor:view',
+                ], $context, $hash->userid))) {
             send_stored_file($file, null, 0, true);
         }
     }